% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_html.r
\name{save_html}
\alias{save_html}
\title{Save http text to file}
\usage{
save_html(html_text, filename = "./page_output/page_output.html")
}
\arguments{
\item{html_text}{http text}

\item{filename}{File name to save}
}
\description{
Save http text to file
}
\examples{
res <- httr::GET("http://google.com/")
html_text <- httr::content(res, "text")
save_html(html_text)

}

